crawl4ai-server-railway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
crawl4ai==0.6.3
groq==0.4.1
python-dotenv==1.0.0
playwright==1.40.0
